+++ server/component/style/resetPassword/tpl_reset_anonymous_user.php
@@ -1,0 +1,11 @@
+<?php
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
+?>
+<div class="container mt-3">
+    <div class="jumbotron">
+        <?php echo $this->text; ?>
+        <?php echo $this->output_reset_anonymous_user(); ?>
+    </div>
+</div>